cast StandardDeviation methods to Double

+++ src/main/java/com/datahub/Datahubtestserver/dataValidation/StandardDeviation.java
@@ -8,9 +8,9 @@ import java.util.stream.Stream;
 public class StandardDeviation {
 
 
-    public static List<Integer> getBadDataIndexes(List<Float> dataList){
-        float mean = calculateMean(dataList);
-        float standardDeviation = calculateStandardDeviation(dataList,mean);
+    public static List<Integer> getBadDataIndexes(List<Double> dataList){
+        double mean = calculateMean(dataList);
+        double standardDeviation = calculateStandardDeviation(dataList,mean);
         return Stream.iterate(0, n -> n+1).
                 limit(dataList.size()).
                 filter(x -> !(Math.abs(dataList.get(x) - mean) < standardDeviation)).
@@ -18,17 +18,17 @@ public class StandardDeviation {
     }
 
 
-    private static float calculateStandardDeviation(List<Float> dataList,Float mean){
+    private static float calculateStandardDeviation(List<Double> dataList,Double mean){
         return (float) Math.sqrt(
                 dataList.stream().
                         map(x -> (x - mean)*(x - mean)).
-                        reduce(0f, Float::sum) / dataList.size()
+                        reduce((double) 0, Double::sum) / dataList.size()
         );
     }
 
 
-    private static float calculateMean(List<Float> dataList){
+    private static double calculateMean(List<Double> dataList){
         return dataList.stream().
-                reduce(0f, Float::sum) / dataList.size();
+                reduce((double) 0, Double::sum) / dataList.size();
     }
 }
